job/requirements.txt
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.2

